+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change history for ui-inventory
 
+## IN PROGRESS
+
+* Save Holdings UUIDs in the Inventory search result. Refs UIIN-1662.
 ## [7.1.4](https://github.com/folio-org/ui-inventory/tree/v8.0.0) (2021-10-05)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v7.1.4...v8.0.0)